+++ src/pages/AddCard/AddCard.tsx
@@ -2,11 +2,9 @@ import { Card } from '@/components/Common';
 import AddCardForm from '@/components/Form/AddCardForm';
 import Layout from '@/components/Layout';
 import { useCardForm } from '@/context/CardFormContext';
-import { useMemo } from 'react';
 
 function AddCard() {
-  const card = useCardForm();
-  const { cardNumber1, cardNumber2, cardNumber3, cardNumber4, year, month, cardOwner } = useMemo(() => card, [card]);
+  const { cardNumber1, cardNumber2, cardNumber3, cardNumber4, year, month, cardOwner } = useCardForm();
 
   return (
     <Layout headerTitle="카드 추가" goBack="/">
