+++ src/pages/Complete/components/index.ts
@@ -1,0 +1,1 @@
+export { default as CompleteForm } from './CompleteForm';
