+++ src/pages/Complete/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useLocationState } from './useLocationState';
